+++ kernel/src/include/bindings/wrapper.h
@@ -46,3 +46,4 @@
 #include <sched/sched.h>
 #include <smp/smp.h>
 #include <time/sleep.h>
+#include <time/clocksource.h>
