+++ kernel/src/time/clocksource.rs
@@ -1,4 +1,4 @@
-use core::intrinsics::log2f32;
+use core::{ffi::c_void, intrinsics::log2f32};
 
 use alloc::{
     boxed::Box,
@@ -12,6 +12,7 @@ use lazy_static::__Deref;
 use crate::{kdebug, kinfo, libs::spinlock::SpinLock, syscall::SystemError};
 
 use super::{
+    jiffies::{ClocksourceJiffies, InnerJiffies},
     timer::{clock, Timer, TimerFunction},
     NSEC_PER_SEC,
 };
@@ -31,12 +32,7 @@ lazy_static! {
     pub static ref DEFAULT_CLOCK:Arc<ClocksourceJiffies> = ClocksourceJiffies::new();
 
 }
-//一些应该放在jiffies里里面的常量 暂时先放一下
-pub const CLOCK_TICK_RATE: u32 = HZ as u32 * 100000;
-pub const JIFFIES_SHIFT: u32 = 8;
-pub const LATCH: u32 = ((CLOCK_TICK_RATE + (HZ as u32) / 2) / HZ as u32) as u32;
-pub const ACTHZ: u32 = sh_div(CLOCK_TICK_RATE, LATCH, 8);
-pub const NSEC_PER_JIFFY: u32 = ((NSEC_PER_SEC << 8) / ACTHZ) as u32;
+
 
 /// 正在被使用时钟源
 pub static CUR_CLOCKSOURCE: SpinLock<Option<Arc<dyn Clocksource>>> = SpinLock::new(None);
@@ -212,6 +208,7 @@ impl dyn Clocksource {
         self.clocksource_enqueue();
         self.clocksource_enqueue_watchdog();
         clocksource_select();
+        kdebug!("clocksource_register successfully");
         return Ok(());
     }
     /// 将时间源插入时间源队列
@@ -373,7 +370,7 @@ impl dyn Clocksource {
             .remove(ClocksourceFlags::CLOCK_SOURCE_WATCHDOG);
         self.update_clocksource_data(cs_data);
         // TODO 停止当前的watchdog
-        // clocksource_stop_watchdog
+        clocksource_stop_watchdog();
     }
 
     /// 将时间源从链表中弹出
@@ -514,7 +511,7 @@ pub fn clocksource_watchdog() -> Result<(), SystemError> {
             .contains(ClocksourceFlags::CLOCK_SOURCE_UNSTABLE)
         {
             // TODO 启动wd thread
-            // BUG 不会重复启动多次嘛
+            run_watchdog_kthread();
 
             continue;
         }
@@ -616,7 +613,6 @@ pub fn clocksource_reset_watchdog() {
 pub fn clocksource_stop_watchdog() {
     let wd = &mut CLOCKSOUCE_WATCHDOG.lock();
     let wd_list = &WATCHDOG_LIST.lock();
-    // BUG 后面的条件不是很懂
     if !wd.is_running || (wd.watchdog.is_some() && !wd_list.is_empty()) {
         return;
     }
@@ -669,62 +665,7 @@ pub fn clocksource_select() {
         // TODO 通知timerkeeping 切换了时间源
     }
 }
-// TODO 应该放在jiffies.rs
-pub const fn sh_div(nom: u32, den: u32, lsh: u32) -> u32 {
-    (((nom) / (den)) << (lsh)) + ((((nom) % (den)) << (lsh)) + (den) / 2) / (den)
-}
-pub struct ClocksourceJiffies(SpinLock<InnerJiffies>);
-pub struct InnerJiffies {
-    data: ClocksourceData,
-    self_ref: Weak<ClocksourceJiffies>,
-}
 
-impl Clocksource for ClocksourceJiffies {
-    fn read(&self) -> CycleNum {
-        CycleNum(clock())
-    }
-
-    fn clocksource_data(&self) -> ClocksourceData {
-        let inner = self.0.lock();
-        return inner.data.clone();
-    }
-
-    fn clocksource(&self) -> Arc<dyn Clocksource> {
-        self.0.lock().self_ref.upgrade().unwrap()
-    }
-    fn update_clocksource_data(&self, _data: ClocksourceData) -> Result<(), SystemError> {
-        let d = &mut self.0.lock().data;
-        d.set_flags(_data.flags);
-        d.set_mask(_data.mask);
-        d.set_max_idle_ns(_data.max_idle_ns);
-        d.set_mult(_data.mult);
-        d.set_name(_data.name);
-        d.set_rating(_data.rating);
-        d.set_shift(_data.shift);
-        return Ok(());
-    }
-}
-impl ClocksourceJiffies {
-    pub fn new() -> Arc<Self> {
-        let data = ClocksourceData {
-            name: "jiffies".to_string(),
-            rating: 1,
-            mask: ClocksourceMask { bits: 0xffffffff },
-            mult: NSEC_PER_JIFFY << JIFFIES_SHIFT,
-            shift: JIFFIES_SHIFT,
-            max_idle_ns: Default::default(),
-            flags: ClocksourceFlags { bits: 0 },
-            watchdog_last: CycleNum(0),
-        };
-        let jieffies = Arc::new(ClocksourceJiffies(SpinLock::new(InnerJiffies {
-            data: data,
-            self_ref: Default::default(),
-        })));
-        jieffies.0.lock().self_ref = Arc::downgrade(&jieffies);
-
-        return jieffies;
-    }
-}
 pub fn clocksource_default_clock() -> Arc<ClocksourceJiffies> {
     DEFAULT_CLOCK.clone()
 }
@@ -735,4 +676,14 @@ pub fn clocksource_init() {
     *FINISHED_BOOTING.lock() = true;
     // 清除不稳定的时钟源
     clocksource_watchdog_kthread();
+}
+
+#[no_mangle]
+pub extern "C" fn rs_clocksource_init() {
+    clocksource_init();
+}
+#[no_mangle]
+pub extern "C" fn rs_clocksource_watchdog_kthread(_data: c_void) -> i32 {
+    clocksource_watchdog_kthread();
+    return 0;
 }
