--- kernel/src/time/hrtimer.rs
@@ -1,1 +1,0 @@
-
