+++ kernel/src/time/jiffies.rs
@@ -1,0 +1,63 @@
+
+//一些应该放在jiffies里里面的常量 暂时先放一下
+pub const CLOCK_TICK_RATE: u32 = HZ as u32 * 100000;
+pub const JIFFIES_SHIFT: u32 = 8;
+pub const LATCH: u32 = ((CLOCK_TICK_RATE + (HZ as u32) / 2) / HZ as u32) as u32;
+pub const ACTHZ: u32 = sh_div(CLOCK_TICK_RATE, LATCH, 8);
+pub const NSEC_PER_JIFFY: u32 = ((NSEC_PER_SEC << 8) / ACTHZ) as u32;
+// TODO 应该放在jiffies.rs
+pub const fn sh_div(nom: u32, den: u32, lsh: u32) -> u32 {
+    (((nom) / (den)) << (lsh)) + ((((nom) % (den)) << (lsh)) + (den) / 2) / (den)
+}
+pub struct ClocksourceJiffies(SpinLock<InnerJiffies>);
+pub struct InnerJiffies {
+    data: ClocksourceData,
+    self_ref: Weak<ClocksourceJiffies>,
+}
+
+impl Clocksource for ClocksourceJiffies {
+    fn read(&self) -> CycleNum {
+        CycleNum(clock())
+    }
+
+    fn clocksource_data(&self) -> ClocksourceData {
+        let inner = self.0.lock();
+        return inner.data.clone();
+    }
+
+    fn clocksource(&self) -> Arc<dyn Clocksource> {
+        self.0.lock().self_ref.upgrade().unwrap()
+    }
+    fn update_clocksource_data(&self, _data: ClocksourceData) -> Result<(), SystemError> {
+        let d = &mut self.0.lock().data;
+        d.set_flags(_data.flags);
+        d.set_mask(_data.mask);
+        d.set_max_idle_ns(_data.max_idle_ns);
+        d.set_mult(_data.mult);
+        d.set_name(_data.name);
+        d.set_rating(_data.rating);
+        d.set_shift(_data.shift);
+        return Ok(());
+    }
+}
+impl ClocksourceJiffies {
+    pub fn new() -> Arc<Self> {
+        let data = ClocksourceData {
+            name: "jiffies".to_string(),
+            rating: 1,
+            mask: ClocksourceMask { bits: 0xffffffff },
+            mult: NSEC_PER_JIFFY << JIFFIES_SHIFT,
+            shift: JIFFIES_SHIFT,
+            max_idle_ns: Default::default(),
+            flags: ClocksourceFlags { bits: 0 },
+            watchdog_last: CycleNum(0),
+        };
+        let jieffies = Arc::new(ClocksourceJiffies(SpinLock::new(InnerJiffies {
+            data: data,
+            self_ref: Default::default(),
+        })));
+        jieffies.0.lock().self_ref = Arc::downgrade(&jieffies);
+
+        return jieffies;
+    }
+}
