+++ kernel/src/time/mod.rs
@@ -3,7 +3,7 @@ use core::{fmt, ops};
 use self::timekeep::ktime_get_real_ns;
 
 pub mod clocksource;
-pub mod hrtimer;
+pub mod jiffies;
 pub mod posix_timer;
 pub mod sleep;
 pub mod timekeep;
