+++ kernel/src/time/timer.h
@@ -2,7 +2,7 @@
 
 #include <common/glib.h>
 #include <driver/timers/HPET/HPET.h>
-
+#include <common/kthread.h>
 // 定义LONG_MAX为最大超时时间 - 允许负数
 #define MAX_TIMEOUT (int64_t)((1ul << 63) - 1)
